Recuperation de l'email utilisateur via le token dans client-ui afin de l'envoyer a la vue books et de pouvoir lier l'email utilisateur a la reservation lors de sa creation. Cet email sert ensuite a envoyé les mails dans le microservice batch. Ajout du batch des reservations avec liste d'attente, priorité par date, creation d'un emprunt, changement des status... Il reste des regles de gestion a finir (en cours).

+++ books/src/main/java/fr/axelallain/books/controller/BooksController.java
@@ -34,7 +34,7 @@ public class BooksController {
     }
 
     @GetMapping("/books/{id}")
-    public Optional<Book> booksById(@PathVariable int id, HttpServletResponse response) {
+    public Optional<Book> booksById(@PathVariable Long id, HttpServletResponse response) {
 
         Optional<Book> book = booksDao.findById(id);
 
@@ -58,7 +58,7 @@ public class BooksController {
     }
 
     @PutMapping("/books/{id}")
-    public void booksEdit(@RequestBody Book book, @PathVariable int id, HttpServletResponse response) {
+    public void booksEdit(@RequestBody Book book, @PathVariable Long id, HttpServletResponse response) {
 
         Optional<Book> bookOptional = booksDao.findById(id);
 
@@ -72,7 +72,7 @@ public class BooksController {
     }
 
     @DeleteMapping("/books/{id}")
-    public void booksDelete(@PathVariable int id) {
+    public void booksDelete(@PathVariable Long id) {
 
         booksDao.deleteById(id);
     }
